requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
openai>=0.27.0
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0
